+++ tests/unit/test_partial_rerun_merge_app.py
@@ -1,5 +1,5 @@
 """Module contains functional requirements for partial rerun merge event handler."""
-
+from copy import deepcopy
 from uuid import uuid4
 
 import pytest
@@ -121,3 +121,45 @@ def test_merge_event_handler_called_from_scheduled_app_calls_needed_handler(cont
 
     partial_merge_app_response = scheduled_app(RAW_EVENT, context)[0]
     assert partial_merge_app_response == expected_response
+
+
+def test_event_parsing_not_failing_on_missing_field(context):
+    """
+    When field is missing we're still able to parse the event and invoke correct
+    function
+    """
+
+    @corva.scheduled(merge_events=True)
+    def scheduled_app(event, api, cache):
+        pytest.fail("Scheduled app was unexpectedly called!")
+
+    @corva.partial_rerun_merge
+    def partial_rerun_merge_app(event, api, asset_cache, rerun_asset_cache):
+        return True
+
+    event_with_missing_parameter: dict = deepcopy(RAW_EVENT)
+    event_with_missing_parameter["data"]["end"] = None
+    partial_merge_app_response = scheduled_app(event_with_missing_parameter, context)[0]
+    assert partial_merge_app_response is True
+
+
+def test_merge_events_does_not_fail_for_partial_rerun_merge_events(context, mocker):
+    """
+    partial rerun merge events shouldn't fail to be processed when used in combination
+    with merge_events parameter
+    """
+
+    @corva.scheduled(merge_events=True)
+    def scheduled_app(event, api, cache):
+        pytest.fail("Scheduled app was unexpectedly called!")
+
+    @corva.partial_rerun_merge
+    def partial_rerun_merge_app(event, api, asset_cache, rerun_asset_cache):
+        return True
+
+    partial_merge_app_response = scheduled_app(RAW_EVENT, context)[0]
+    assert partial_merge_app_response is True
+
+
+def test_partial_rerun_decorator_can_be_called_directly(context):
+    corva.partial_rerun_merge()